feat: implement subscription apis: cancel, list by deviceNo

+++ src/test/java/com/github/lucasdengcn/billing/api/SubscriptionControllerIntegrationTest.java
@@ -622,7 +622,7 @@ class SubscriptionControllerIntegrationTest {
                 .andExpect(status().isOk())
                 .andExpect(jsonPath("$").isArray())
                 .andExpect(jsonPath("$", hasSize(2)))
-                .andExpect(jsonPath("$[*].deviceId").value(everyItem(equalTo(testDevice.getId()))));
+                .andExpect(jsonPath("$[*].deviceId").value(everyItem(equalTo(testDevice.getId().intValue()))));
     }
 
     @Test
